update types [skip ci]
